+++ apps/ced-browser-be/CHANGELOG.md
@@ -1,5 +1,13 @@
 # ced-browser-be
 
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies [cc58ed4]
+  - @pagopa/io-core-adapter-fastify@0.0.8
+  - @pagopa/io-core-adapter-fims@0.0.12
+
 ## 0.1.15
 
 ### Patch Changes